+++ ToyManagementProject/ToyManagementProject.Services/ToyService.cs
@@ -9,15 +9,15 @@ using ToyManagementProject.Domain.Interfaces.Repositories;
 
 namespace ToyManagementProject.Services
 {
-	public class ToyService: IToyService
-	{		
+	public class ToyService : IToyService
+	{
 		private readonly IToyRepository _toyRepository;
 		private readonly IUnitOfWork _uow;
 		private readonly IMapper _mapper;
 		//private readonly IValidator<Toy> _toyValidator;
-		public ToyService(IToyRepository toyRepository, 
-						  IUnitOfWork uow, 
-						  IMapper mapper, 
+		public ToyService(IToyRepository toyRepository,
+						  IUnitOfWork uow,
+						  IMapper mapper,
 						  IValidator<Toy> toyValidator)
 		{
 			_toyRepository = toyRepository;
@@ -29,17 +29,13 @@ namespace ToyManagementProject.Services
 		{
 			try
 			{
-				if (!toy.IsValid()) 
+				if (!toy.IsValid())
 				{
 					return Result<ToyDto>.Failure(toy.ErrorsNotifications);
-				}
+				}		
 
-				//var validateErrors = _toyValidator.Validate(toy);
-
-				//if (validateErrors.Any())
-				//	return Result<ToyDto>.Failure(validateErrors);
-											
 				await _toyRepository.AddAsync(toy);
+				
 				await _uow.CommitAsync();
 
 				return Result<ToyDto>.Success(_mapper.Map<ToyDto>(toy));
@@ -47,27 +43,32 @@ namespace ToyManagementProject.Services
 			catch (Exception ex)
 			{
 				return Result<ToyDto>.Failure(new List<string> { $"Error AddAsync: {ex.Message}" });
-			}		
+			}
 		}
 
 		public async Task<Result<ToyDto>> DeleteAsync(int id)
 		{
-			try
-			{
-				var toy = _toyRepository.GetByIdAsync(id);
-				
-				if (toy == null) 
-				{
-					return Result<ToyDto>.Failure($"Toy doesn´t exists");
-				}
+			var toy = await _toyRepository.GetByIdAsync(id);
 
+			if (toy == null)
+			{
+				return Result<ToyDto>.Failure($"Toy doesn´t exists");
+			}
+
+			if (!toy.IsValid())
+			{
+				return Result<ToyDto>.Failure($"{toy.ErrorsNotifications}");
+			}
+
+			try
+			{			
 				await _toyRepository.DeleteAsync(id);
-				
+
 				await _uow.CommitAsync();
 
 				var toyDTO = _mapper.Map<ToyDto>(toy);
 
-				return Result<ToyDto>.Success(toyDTO);
+				return Result<ToyDto>.Success(null, "Toy deleted");
 			}
 			catch (Exception ex)
 			{
@@ -78,46 +79,50 @@ namespace ToyManagementProject.Services
 		public async Task<Result<ToyDto>> GetByIdAsync(int id)
 		{
 			var toy = await _toyRepository.GetByIdAsync(id);
+
 			if (toy == null)
 			{
 				return Result<ToyDto>.Failure($"Toy is not registered.");
 			}
 
+			if (!toy.IsValid())
+			{
+				return Result<ToyDto>.Failure($"{toy.ErrorsNotifications}");
+			}
+
 			return Result<ToyDto>.Success(_mapper.Map<ToyDto>(toy));
-		}	
+		}
 		public async Task<Result<ToyDto>> UpdateAsync(Toy toy)
-		{						
-			if (toy.ErrorsNotifications.Any())
-            {
+		{
+			if (!toy.IsValid())
+			{
 				return Result<ToyDto>.Failure(toy.ErrorsNotifications);
 			}
 
-            try
-			{				
+			try
+			{
 				await _toyRepository.UpdateAsync(toy);
+
 				await _uow.CommitAsync();
 
-				return Result<ToyDto>.Success(_mapper.Map<ToyDto>(toy));
+				return Result<ToyDto>.Success(null, "Toy updated");
 			}
 			catch (Exception ex)
 			{
 				return Result<ToyDto>.Failure(new List<string> { $"Error UpdateAsync: {ex.Message}" });
 				throw;
 			}
-			
 		}
 
 		public async Task<Result<IEnumerable<ToyDto>>> GetAllAsync()
 		{
 			var toys = await _toyRepository.GetAllAsync();
-			if (toys == null)
+
+			if (toys == null || toys.Count == 0)
 			{
 				return Result<IEnumerable<ToyDto>>.Failure($"toys is empty");
 			}
-
 			return Result<IEnumerable<ToyDto>>.Success(_mapper.Map<IEnumerable<ToyDto>>(toys));
-
-
 		}
 	}
 }
